+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 🐛 Bug Fix
 
+- Fix an issue with some BetterTTV emotes.
 - Improve error handling for unavailable channels.
 
 # 0.1.0